+++ Services/ProjectMan/V4/Model/UpdateIssueV4Response.cs
@@ -139,6 +139,12 @@ namespace HuaweiCloud.SDK.ProjectMan.V4.Model
         /// <summary>
         /// 
         /// </summary>
+        [JsonProperty("order", NullValueHandling = NullValueHandling.Ignore)]
+        public IssueResponseV4Order Order { get; set; }
+
+        /// <summary>
+        /// 
+        /// </summary>
         [JsonProperty("severity", NullValueHandling = NullValueHandling.Ignore)]
         public IssueItemSfV4Severity Severity { get; set; }
 
@@ -147,6 +153,24 @@ namespace HuaweiCloud.SDK.ProjectMan.V4.Model
         /// </summary>
         [JsonProperty("status", NullValueHandling = NullValueHandling.Ignore)]
         public IssueItemSfV4Status Status { get; set; }
+
+        /// <summary>
+        /// 工作项发布版本号
+        /// </summary>
+        [JsonProperty("release_dev", NullValueHandling = NullValueHandling.Ignore)]
+        public string ReleaseDev { get; set; }
+
+        /// <summary>
+        /// 缺陷发现版本号（仅Bug类型工作项具备该字段）
+        /// </summary>
+        [JsonProperty("find_release_dev", NullValueHandling = NullValueHandling.Ignore)]
+        public string FindReleaseDev { get; set; }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        [JsonProperty("env", NullValueHandling = NullValueHandling.Ignore)]
+        public IssueResponseV4Env Env { get; set; }
 
         /// <summary>
         /// 
@@ -195,8 +219,12 @@ namespace HuaweiCloud.SDK.ProjectMan.V4.Model
             sb.Append("  name: ").Append(Name).Append("\n");
             sb.Append("  parentIssue: ").Append(ParentIssue).Append("\n");
             sb.Append("  priority: ").Append(Priority).Append("\n");
+            sb.Append("  order: ").Append(Order).Append("\n");
             sb.Append("  severity: ").Append(Severity).Append("\n");
             sb.Append("  status: ").Append(Status).Append("\n");
+            sb.Append("  releaseDev: ").Append(ReleaseDev).Append("\n");
+            sb.Append("  findReleaseDev: ").Append(FindReleaseDev).Append("\n");
+            sb.Append("  env: ").Append(Env).Append("\n");
             sb.Append("  tracker: ").Append(Tracker).Append("\n");
             sb.Append("  updatedTime: ").Append(UpdatedTime).Append("\n");
             sb.Append("  closedTime: ").Append(ClosedTime).Append("\n");
@@ -325,6 +353,11 @@ namespace HuaweiCloud.SDK.ProjectMan.V4.Model
                     this.Priority.Equals(input.Priority))
                 ) && 
                 (
+                    this.Order == input.Order ||
+                    (this.Order != null &&
+                    this.Order.Equals(input.Order))
+                ) && 
+                (
                     this.Severity == input.Severity ||
                     (this.Severity != null &&
                     this.Severity.Equals(input.Severity))
@@ -333,6 +366,21 @@ namespace HuaweiCloud.SDK.ProjectMan.V4.Model
                     this.Status == input.Status ||
                     (this.Status != null &&
                     this.Status.Equals(input.Status))
+                ) && 
+                (
+                    this.ReleaseDev == input.ReleaseDev ||
+                    (this.ReleaseDev != null &&
+                    this.ReleaseDev.Equals(input.ReleaseDev))
+                ) && 
+                (
+                    this.FindReleaseDev == input.FindReleaseDev ||
+                    (this.FindReleaseDev != null &&
+                    this.FindReleaseDev.Equals(input.FindReleaseDev))
+                ) && 
+                (
+                    this.Env == input.Env ||
+                    (this.Env != null &&
+                    this.Env.Equals(input.Env))
                 ) && 
                 (
                     this.Tracker == input.Tracker ||
@@ -399,10 +447,18 @@ namespace HuaweiCloud.SDK.ProjectMan.V4.Model
                     hashCode = hashCode * 59 + this.ParentIssue.GetHashCode();
                 if (this.Priority != null)
                     hashCode = hashCode * 59 + this.Priority.GetHashCode();
+                if (this.Order != null)
+                    hashCode = hashCode * 59 + this.Order.GetHashCode();
                 if (this.Severity != null)
                     hashCode = hashCode * 59 + this.Severity.GetHashCode();
                 if (this.Status != null)
                     hashCode = hashCode * 59 + this.Status.GetHashCode();
+                if (this.ReleaseDev != null)
+                    hashCode = hashCode * 59 + this.ReleaseDev.GetHashCode();
+                if (this.FindReleaseDev != null)
+                    hashCode = hashCode * 59 + this.FindReleaseDev.GetHashCode();
+                if (this.Env != null)
+                    hashCode = hashCode * 59 + this.Env.GetHashCode();
                 if (this.Tracker != null)
                     hashCode = hashCode * 59 + this.Tracker.GetHashCode();
                 if (this.UpdatedTime != null)